requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
deepface==0.0.95
fastapi==0.116.1
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.74.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.11.1
libclang==18.1.1
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.3
mtcnn==1.0.0
namex==0.1.0
numpy==2.1.3
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==5.29.5
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
requests==2.32.4
retina-face==0.0.17
rich==14.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
tf_keras==2.19.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
